Move to example separate folder
