+++ src/WebApp.Core/app/components/bluetooth/bluetooth-page.html
