+++ src/WebApp.Core/app/startup.ts
@@ -11,7 +11,7 @@ import { NavMenuViewModel } from "./components/nav-menu";
 import { WishlistViewModel } from "./components/wishlist-page";
 import { MyOrdersViewModel } from "./components/myorders-page";
 import { NfcViewModel } from "./components/nfc-page";
-import { BlueToothViewModel } from "./components/bluetooth-page";
+import { BlueToothViewModel } from "./components/bluetooth/bluetooth-page";
 import * as $ from "jquery";
 import { PaymentRequestPageViewModel } from "./components/payment-request-page";
 
@@ -66,7 +66,7 @@ ko.components.register("payment-request-page", {
 });
 ko.components.register("bluetooth-page", {
   viewModel: BlueToothViewModel,
-  template: { require: "text!components/bluetooth-page.html" }
+  template: { require: "text!components/bluetooth/bluetooth-page.html" }
 });
 
 // Start our app
